refactor: Refactor code
