+++ main.py
@@ -1,14 +1,12 @@
 from account import Account
-from csvManager import CSVManager
+from accountManager import AccountManager
 
 
 def main():
     acc = Account('test', 'ozef', 'No website')
-    print(acc)
-    csv_manager = CSVManager()
-    print(csv_manager)
-    csv_manager.add_account(acc)
-    accounts = csv_manager.get_all_accounts()
+    account_manager = AccountManager()
+    account_manager.add_account(acc)
+    accounts = account_manager.get_all_accounts()
     [print(acc) for acc in accounts]
